fix txns banners
